+++ src/main/scala/com/bharatsim/model/Office.scala
@@ -2,7 +2,7 @@ package com.bharatsim.model
 
 import com.bharatsim.engine.models.Network
 
-case class Office(id: Int) extends Network {
+case class Office(id: Long) extends Network {
   addRelation[Person]("EMPLOYER_OF")
 
   override def getContactProbability(): Double = 1
