+++ src/main/scala/com/bharatsim/model/School.scala
@@ -2,7 +2,7 @@ package com.bharatsim.model
 
 import com.bharatsim.engine.models.Network
 
-case class School(id: Int) extends Network {
+case class School(id: Long) extends Network {
   addRelation[Person]("TEACHES")
 
   override def getContactProbability(): Double = 1
